renamed Blogs to blogs, resolved conflicts in constants, navigation.
