+++ constants/backgrounds.js
@@ -1,0 +1,22 @@
+const football = require("../assets/images/football.jpg")
+const tennis = require("../assets/images/tennis.jpg")
+const library = require("../assets/images/library.jpg")
+const runner = require("../assets/images/relay-runner.jpg")
+
+const aurelius = require("../assets/images/Gallery/marcus-aurelius.jpg")
+const muhammadali = require("../assets/images/Gallery/muhammad-ali.jpg")
+const freud = require("../assets/images/Gallery/sigmund-freud.jpg")
+const kobe = require("../assets/images/Gallery/kobe-bryant.jpg")
+const tombrady = require("../assets/images/Gallery/tom-brady.jpg")
+
+export default {
+    football,
+    tennis,
+    library,
+    runner,
+    aurelius,
+    muhammadali,
+    freud,
+    kobe,
+    tombrady,
+}
